RF+ENH: output messages to stderr not stdout + move printing out of "controllers/models"
Most of the tools which operate over some other tool/streams (e.g. pv, datalad
run, etc) strive to not interfer with stdout so then they could be used
within pipe constructs.  If such wrapper tool outputs to stdout, then it could
not be used within a pipe, e.g. in

   duct my_amportant_compute | grep "a"

construct.  The same goes for git and git-annex, datalad, etc, unless it is
explicitly a command intended to display something (e.g. "datalad wtf").

That is why important to not just "print" to stdout but use e.g. stderr which
is not piped through.  Relates also to issue on logging which should default to
stderr not stdout.