ENH: give "typical" shell behavior when command is not found to be executed
both bash and zsh exit then with 127 and message like "command_name: command not found",
so duct now will be have the same instead of throwing the exception traceback
into the user.

Also, there is no reason to announce to the user that we are saving some log files in such
a case. But since we might have created them by our attempt to execute, we will just remove
them before exiting with an error.  This way we would not leave junk around