RF: Add logging, dissolve duct_print (INFO level), add CLI option -l, dissolve --quiet
When we have logging, and it also goes to stderr, there is really no need for
some dedicated "duct_print". Now everything can be controlled with -l and
no additional --quiet is needed.

the only location is when to report that command is incorrect - that is where I used
explicit print to stderr since so far the only location.

I also added a few lgr.debug level statements.