RF+BF: update maxes on each sample, more logging during monitoring
We figured out occasional "slow down" -- it is due to nvidia-smi invocation
during collection of summary. Relevant now issue:
- https://github.com/con/duct/issues/145

We added more logging in the monitoring loop so we would know when/why we
are leaving it.

We also added explicit 2nd break out from the loop if process is gone before
we manage to ps.