TEMP: remove log files if interrupted via Ctrl-C
This is not how it should be!  it would depend on when it is interrupted and
what was the actual process return code.  Immediate usecase is my impatience
with current implementation still waiting whenever command such as  "ls" is
done already.

Then when I interrupt with Ctrl-C we should still dump all the results and
proceed to treat on how we would treat if command failed or not.  It would also
relate to

- https://github.com/con/duct/issues/184

so if Ctrl-C the child process -- it is non-0 exit, and we would treat as
"failed run".

So we should add it more smartly than what I did here