feat: Pass Ctrl+c interrupt to executed command
Uses multiprocessing, not multithreading Python signal handler
registration only works in the main thread. Also, we dont want to test
sending the SIGINT to the command, we want to test sending SIGINT to
duct, which should pass it through to the command.

We have to set the wait time pretty high for this to pass CI tests.
Specifically, when using pypy and coverage at the same time, we need
extra time between SIGINTs.